SPDBT-4364
